Added method to pull all unstarted public games. Added a lobby page. Moved invitation-related JDBCs/DAOs to their own files (previously in the Game section). Open games are displaying in lobby page, sorted by number of players. Added categories to invitation/open game windows.
