+++ src/main/webapp/WEB-INF/jsp/common/invitations.jsp
@@ -1,5 +1,6 @@
 <%@ taglib uri="http://java.sun.com/jsp/jstl/core" prefix="c" %>
 <%@ taglib prefix="form" uri="http://www.springframework.org/tags/form" %>
+<%@ taglib prefix="fmt" uri="http://java.sun.com/jsp/jstl/fmt" %>
 
 <c:url var="joinGameURL" value="/joinGame"/>
 <c:url var="rejectGameURL" value="/rejectGame"/>
@@ -10,7 +11,11 @@
 		<div class="invites__invite u-margin-bottom-small">
 			<div class="invites__details u-margin-bottom-small">
 				Game Code: <c:out value="${invitation.gameCode }"/><br>
-				Invited By: <c:out value="${invitation.invitedBy }"/>
+				Invited By: <c:out value="${invitation.invitedBy }"/><br>
+				Categories: ${pairs.get(invitation).get(0).categoryName}<c:forEach var="index" begin="1" end="${pairs.get(invitation).size() - 1}">,
+				<fmt:parseNumber var="intValue" value="${index}" integerOnly="true"/>
+				${pairs.get(invitation).get(intValue).categoryName}
+				</c:forEach>
 			</div>
 			<div class="invites__buttons">
 				<form method="POST" action="${joinGameURL}">
@@ -18,7 +23,6 @@
 					<input type="hidden" value="${invitation.gameCode }" name="gameCode"/>
 				</form>
 				
-<!-- 			JEFF: THIS BUTTON NEEDS CONFIGURING -->
 				<form method="POST" action="${rejectGameURL}">
 					<input type="submit" class="button--small" value="Reject"/>
 					<input type="hidden" value="${invitation.gameCode }" name="gameCode"/>
